Move UDDataset and associated classes to new file
Also don't create a UDDataset directly in client code; instead, use
`model.encode_dataset` to create one. All of the config needed for the
dataset creation is already in the model, so this way clients don't have
to worry about copying fields between the classes.
